% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeConceptRepresentation.R
\name{computeConceptRepresentation}
\alias{computeConceptRepresentation}
\title{Computes vector representations of a (set of) concept(s) based on a given
model, and the terms in the model vocabulary that define that concept(s)
(similar to the approaches described in Garten et al. 2018, BRM).}
\usage{
computeConceptRepresentation(model, term_set = NULL)
}
\arguments{
\item{model}{matrix containing model loaded using loadModel()}

\item{term_set}{a list with one element for each dictionary concept, with 
each element containing a character vector with all terms defining the 
concept, or a row/column-named dataframe with terms in the rows and weights
in the columns}
}
\value{
a concept by dimension matrix containing the vector representation of
  each concept in the dictionary
}
\description{
Computes vector representations of a (set of) concept(s) based on a given
model, and the terms in the model vocabulary that define that concept(s)
(similar to the approaches described in Garten et al. 2018, BRM).
}
