% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeConceptRepresentationWithWildcards.R
\name{computeConceptRepresentationWithWildcards}
\alias{computeConceptRepresentationWithWildcards}
\title{Computes vector representations of a (set of) concept(s) based on a given 
model, and the terms in the model vocabulary that define that concept(s) 
(similar to the approaches described in Garten et al. 2018, BRM).}
\usage{
computeConceptRepresentationWithWildcards(model, term_set,
  remove_missing = FALSE)
}
\arguments{
\item{model}{a term by dimension matrix containing model loaded using 
loadModel()}

\item{term_set}{a dataframe with the columns "concept" (naming the dictionary
concepts), "root" (naming the word stems used to group the terms), "term"
(listing the terms to be used to construct the concept representations),
and "include" (optional logical variable denoting whether a given term
should be included in the concept representation)}
}
\value{
a concept by dimension matrix containing the vector representation of
  each concept in the dictionary
}
\description{
Computes vector representations of a (set of) concept(s) based on a given 
model, and the terms in the model vocabulary that define that concept(s) 
(similar to the approaches described in Garten et al. 2018, BRM).
}
