% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMatrixSimilarity.R
\name{computeMatrixSimilarity}
\alias{computeMatrixSimilarity}
\title{Internal function for computing the semantic similarity between two matrices
of vector representations.}
\usage{
computeMatrixSimilarity(x, y)
}
\arguments{
\item{x}{an m*k matrix containing the vector representation of some entities}

\item{y}{an n*k matrix containing the vector representation of some entities}
}
\value{
an m*k matrix of cosine similarities between the rows of x and y
}
\description{
Internal function for computing the semantic similarity between two matrices
of vector representations.
}
