% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTextVectors.R
\name{computeTextVectors}
\alias{computeTextVectors}
\title{Computes vector representations of a (set of) text(s) based on a given
model.}
\usage{
computeTextVectors(text_vec, model)
}
\arguments{
\item{text_vec}{character vector of texts to process}

\item{model}{matrix containing GloVe model loaded using loadModel()}
}
\value{
a text by dimension matrix containing the vector representation of
  each text
}
\description{
Computes vector representations of a (set of) text(s) based on a given
model.
}
