% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTextToGloveFormat.R
\name{convertTextToGloveFormat}
\alias{convertTextToGloveFormat}
\title{Converts text to a format consistent with the GloVe model vocabulary.}
\usage{
convertTextToGloveFormat(text_vec)
}
\arguments{
\item{text_vec}{character vector of texts to process}
}
\value{
character vector of processed texts
}
\description{
Converts text to a format consistent with the GloVe model vocabulary.
}
