% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTopItems.R
\name{findTopItems}
\alias{findTopItems}
\title{Function to locate the most (dis)similar items to a concept.}
\usage{
findTopItems(concept_similarity, decreasing = TRUE, n = 10)
}
\arguments{
\item{concept_similarity}{an item by concept matrix}

\item{decreasing}{logical denoting whether to search for most (TRUE) vs.
least (FALSE) similar items}

\item{n}{integer denoting the number of most (dis)similar items to find for
each concept}
}
\value{
a list of the most (dis)similar items for each concept, with one
  element for each concept, containing a named vector of similarities
}
\description{
Function to locate the most (dis)similar items to a concept.
}
