% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTopTerms.R
\name{findTopTerms}
\alias{findTopTerms}
\title{Locates the most (dis)similar terms in a vocabulary to a set of concepts.}
\usage{
findTopTerms(term_concept_similarity, decreasing = TRUE, n = 10)
}
\arguments{
\item{term_concept_similarity}{a term by concept matrix computed using 
computeTermConceptSimilarity()}

\item{decreasing}{logical denoting whether to search for most (TRUE) vs.
least (FALSE) similar terms}

\item{n}{integer denoting the number of most (dis)similar terms to find for
each concept}
}
\value{
a list of the most (dis)similar terms for each concept, with one
  element for each concept, containing a named vector of similarities
}
\description{
Locates the most (dis)similar terms in a vocabulary to a set of concepts.
}
