% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWildcards.R
\name{findWildcards}
\alias{findWildcards}
\title{Perform wildcard search for terms in a model's vocabulary.}
\usage{
findWildcards(x, model)
}
\arguments{
\item{x}{a character vector of words}

\item{model}{matrix containing GloVe model loaded using loadModel()}
}
\value{
a list with one element per item in x, containing wildcard search
  results
}
\description{
Perform wildcard search for terms in a model's vocabulary.
}
