% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModel.R
\name{loadModel}
\alias{loadModel}
\title{Loads a pre-trained word embedding model (e.g., GloVe).}
\usage{
loadModel(file, max_vocab_size = NULL)
}
\arguments{
\item{file}{character with path to model}

\item{max_vocab_size}{optional integer specifying the number of terms
to read}
}
\value{
a term by dimension matrix containing the vector representation of
  each term in the vocabulary
}
\description{
Loads a pre-trained word embedding model (e.g., GloVe).
}
