% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeWhiteSpace.R
\name{removeWhiteSpace}
\alias{removeWhiteSpace}
\title{Removes unnecessary whitespace from a character vector of texts.}
\usage{
removeWhiteSpace(text_vec)
}
\arguments{
\item{text_vec}{character vector of texts to process}
}
\value{
character vector of processed texts
}
\description{
Removes unnecessary whitespace from a character vector of texts.
}
